moment_inverter| complete inversion code multiprocessing over mass bins
